requirements.txt
torch==2.11.0
torchvision==0.26.0
torchaudio==2.11.0
pillow
numpy
typing
scikit-image
huggingface_hub
kornia
timm
transformers>=4.39.1
gradio
ipywidgets==8.1.8
matplotlib==3.10.8
requests==2.33.1
python-dotenv==1.2.2
google-generativeai